Refactor AppPage component and import AppHeader component
