+++ src/shared/pages/AppPage.tsx
@@ -1,19 +1,7 @@
-import {
-	AppBar,
-	Avatar,
-	Container,
-	IconButton,
-	Toolbar,
-	Typography,
-} from '@mui/material';
+import { Container } from '@mui/material';
 import React, { useEffect, useLayoutEffect, useRef, useState } from 'react';
 import { AppConstants } from '../constants/AppConstants';
-import UserMenu from '../../features/user/ui/UserMenu';
-import AppDrawer from '../../common/navigation/AppDrawer';
-import { useNavigate } from 'react-router-dom';
-import { AppIcon, AppIcons } from '../Icon';
-import { useUserMocker } from '../../mockers/userMocker';
-import { Placeholder } from '../components/Placeholder';
+import AppHeader from '../components/AppHeader';
 
 export interface AppPageProps {
 	children: React.ReactNode;
@@ -27,12 +15,6 @@ export const AppPage: React.FC<AppPageProps> = ({
 }) => {
 	const appBarRef = useRef<HTMLDivElement>();
 	const [containerHeight, setContainerHeight] = useState(0);
-	const [userMenuAnchorEl, setUserMenuAnchorEl] = useState<null | HTMLElement>(
-		null
-	);
-	const [mobileOpen, setMobileOpen] = useState(false);
-
-	const navigate = useNavigate();
 
 	useEffect(() => {
 		document.title = title + ' | ' + AppConstants.name;
@@ -44,100 +26,9 @@ export const AppPage: React.FC<AppPageProps> = ({
 		}
 	}, [appBarRef.current?.offsetHeight]);
 
-	const user = useUserMocker();
-
-	const handleDrawerToggle = () => {
-		setMobileOpen((prevState) => !prevState);
-	};
-
-	const handleOpenUserMenu = (event: React.MouseEvent<HTMLElement>) => {
-		setUserMenuAnchorEl(event.currentTarget);
-	};
-
-	const handleCloseUserMenu = () => {
-		setUserMenuAnchorEl(null);
-	};
-
-	const canGoBack = () => {
-		return false; // TODO: Fix this
-	};
-
 	return (
 		<React.Fragment>
-			<div ref={appBarRef as React.RefObject<HTMLDivElement>}>
-				<AppBar
-					position='static'
-					style={{
-						minHeight: '5rem',
-					}}
-				>
-					<Toolbar>
-						{canGoBack() ? (
-							// Back button
-							<IconButton
-								size='large'
-								edge='start'
-								color='inherit'
-								aria-label='back-button'
-								sx={{ mr: 2 }}
-								onClick={() => navigate(-1)}
-							>
-								{AppIcon(AppIcons.chevronLeft)}
-							</IconButton>
-						) : null}
-						{/* Hamburger menu */}
-						<IconButton
-							size='large'
-							edge='start'
-							color='inherit'
-							aria-label='menu'
-							sx={{
-								mr: 2,
-								display: {
-									// sm: 'none'
-								},
-							}}
-							onClick={handleDrawerToggle}
-						>
-							{AppIcon(AppIcons.menu)}
-						</IconButton>
-
-						{/* Title */}
-						{typeof title === 'string' ? (
-							<Typography variant='h5' component='div' sx={{ flexGrow: 1 }}>
-								{title}
-							</Typography>
-						) : (
-							title
-						)}
-
-						{/* Profile & Account */}
-						<div>
-							<IconButton
-								size='small'
-								aria-label='account of current user'
-								aria-controls='menu-appbar'
-								aria-haspopup='true'
-								onClick={handleOpenUserMenu}
-								color='inherit'
-							>
-								<Placeholder loading={!user} variant='circular'>
-									<Avatar
-										alt='User'
-										// src='#'
-										src={user?.avatar}
-									/>
-								</Placeholder>
-							</IconButton>
-							<UserMenu
-								anchorEl={userMenuAnchorEl}
-								handleCloseMenu={handleCloseUserMenu}
-							/>
-						</div>
-					</Toolbar>
-				</AppBar>
-			</div>
-			<AppDrawer open={mobileOpen} onDrawerToggle={handleDrawerToggle} />
+			<AppHeader title={title} appBarRef={appBarRef} />
 			<Container
 				style={{
 					// backgroundColor: 'Background',
